Unify Language Extensions (#261)
### Pull Request Description
Moving extensions into `stack.yaml` means that they are now unified across the codebase. This will ensure that we don't run into extension conflicts during interactive development, and furthermore means that we can keep a better eye on the extensions in the codebase. 

### Checklist
Please include the following checklist in your PR:

- [x] The documentation has been updated if necessary.
- [x] The code conforms to the [Luna Style Guide](https://github.com/luna/wiki/blob/master/code-style/01.general.md).
- [x] The code has been tested where possible.