Parser Attaching Ids to Expressions (#577)
This PR makes two changes:
1. parser will give IDs to nodes' expressions (unless already set);
2. `ASTOf.equals` will not compare IDs

Both were already agreed to and implemented as part of `wip/mwu/graph-syntax` branch (though 1. in quite different shape). However, it was abandoned after Rust was selected as the technology for IDE "backend". Now, as we are reimplementing double representation in Rust we need parts of the technology back in the parser.

Prerequisite for https://github.com/luna/ide/issues/186.
Implements https://github.com/luna/ide/issues/228.