make "CloseButton" not be a View
The component is currently presumed to be an internal widget, and
doesn't have keyboard shortcuts. Also, its use is currently not well
aligned with the claimed "CloseButton" label, as it's also used in
function of a "fullscreen button". The decision can be revisited in
future if a different understanding is reached.