Dynamically replace builtin methods with functions
This is WIP. This change does not attempt to handle @Builtin_Type. Yet.

During the initialization of Builtins we collect info about root nodes
representing generated AST truffle nodes. Such nodes will dynamically
replace implementations of @Builtin_Method stubs during IR to Truffle
translation.
There are still rough edges to work out but checking this in to see if
this is the direction we want to take.