Deal with Polyglot split by selectively importing
Sadly because of the move of the functions to a separate module and
@Builtin_Type Polyglot still being in Builtins it's tricky to get auto
imports that don't require manual adjustments.
The main issue is with `PolyglotBranchNode` in `IrToTruffle` that
requires a reference to the polyglot atom constructor.
Decided to make this incremental step to do some benchmarking of
@Builtin_Method solution and then move on @Builtin_Type.
The latter will require a separate pass, hence the additional step.

Additionally removed unused warnings for methods annotated with
@Builtin_Method.