Move Array type out of Builtins
The change moves another type, Array, out of Builtins.enso.
Also introduced a couple of changes to make it possible:
- Complex Types cannot erase annotations of its atoms
- method lookup algorithm has to take into account situation where
Complex Type, it's Atom *and* module use the same name, which makes name
resolution slightly ambiguous. The new version ensures that we don't
have to use qualified names *every single time*

This change introduces some lookup maps for checking if the given atom
constructor is a builtin type. Using maps (and strings as keys) is not
always necessary or unambiguous, and will work on a more permanent
solution.