Move Any to stdlib
In order to be able to locally extend a builtin type, we have to be able
to then find it during symbol resolution.
Since the symbol is being added to a module in stdlib but resolution
would be using Any special builtin type, the latter has to be aware of
the former via shadowdefinitions reference.