Workaround separate compilation problems
In separate compilation annotation processor will only process files
that were touched. However, since we now create metadata files with bulk
information and it is not possible to append/modify exisiting resource
files, it will simply overwrite the original resource with a fraction
of the original information. This would in turn lead to runtime issues.

The only solution was to clean and compile from scratch which is not
really acceptable in regular dev.
Alternatively, in the past we were creating metadata file per source
file. The latter however meant that during Builtins initialization we
had to do a bit of File traversing.

Instead we will read the previous resource file before overriding it,
update entries that changed and append the past contents. It doesn't
deal with all cases (renaming/delete) but neither would the
alternatives.