Move all error types + text
A number of re-adjustments were necessary since Error/Text types are
tested extensively. Notice one additional test in DataflowErrorsTest
where Syntax Error has a slightly different display text when caught.
This was distilled from RuntimeServerTest, which I think had it wrong
the last time but that needs to be verified.