Move Nothing and Function to stdlib
Nothing is rather present in a number of tests, hence the number of
changes.
Also discovered a bug in the initialization of shadow definitions. By
using the scope from stdlib we were throwing away all synthetic members
that get created as part of AtomConstructor initialization, e.g. getters
for fields. This became apparent only when trying to remove many
builtins imports which were still bringing them into scope.

Also, one of `Any` builtin methods wasn't defined on the type, just as a
node.

Finally, added timings to tests to get a better idead of bottlenecks.