Rename local 'reverse' method in List benchmarks
`reverse` is defined in `List` module and since `List` type was moved to
stdlib, method resolution gives a higher priority to the one defined in
stdlib. Local methods are only looked up when the former fails, which is
clearly not the case now.
We should probably have some warnings for that.