Builtins is dead, long live builtins
The last in the bunch of official builtins moved to stdlib.
Unfortunately couldn't move to a better hierarchy in stdlib such as
- Data/Number.enso
- Data/Number/Integer.enso
- Data/Number/Decimal.enso
because symbol resolution when pattern matching on Number always picks
the Module symbol not the type and changing everything to `case
Number.Number` is super ugly. Had similar problems with other modules
but this one would be nasty to use that way.

Discovered some typos in original builtins (probably due to c&p) which
got only revealed after methods were no longer being hardcoded when
registering them with constructors.
Also, some builtin methods nodes appear more than once under different
names, hence the introduction of aliases parameter.

Lastly, Number and Integer does not conform to the rest of the schema
since their definitions in Builtins.enso were only for documentation
purposes, or at least it appears so. Hence the tiny exception when
translating IR to Truffle to ignore their builtin methods.