Auto-generate Builtins Constants
BuiltinType annotation now provides an additional parameter that stores
a link to the corresponding definition in stdlib.

TypesFromProxy provides a single static method `fromTypeSystem` which converts
from type-system type names to atoms.
It is a proxy because it could easily be placed inside {@link org.enso.interpreter.runtime.builtin.Builtins}.
Except that by the time Builtins is compiled it requires ConstantsGen to be present, which is being generated via
the TypeProcessor at a late stage. Similarly {@link org.enso.interpreter.runtime.type.Types} requires
ConstantsGen to be in the same package which creates a catch-22 situation.