Generate method nodes for varargs
When a method or constructor has vararg parameter, we may want to
generate multiple BuiltinMethod representation, up to some number of
parameters.
The approach has been extended in a generic way to take it into
consideration when `expandVargars` parameter is bigger than zero.

This allowed us to remove 4 more hardcoded classes.