Rewrite DSL, allow for repeated exception wrapping
The existing Builtin DSL was becoming a bit too crowded with multiple
elements and default values. That wouldn't scale so had to refactor it.

The functionality stays almost the same except that now we really can
have repeatable @WrapException annotation indicating multiple
catch-clauses for possible exceptions that need to be wrapped.

Added a ton of documentation as the implenetation of the processor is
getting a bit more complicated.