Fix a problem in separate compilation
Discovered a problem with separate compilation when generating exception
wrappers. The latter requires information about all builtin types
annotations but that is not present in separate compilation.
Instead we do a similar trick as in other processors - we read metadata
and diff the results.

The change is bigger than a single method because I didn't want to write
a yet another parser of metadata entries. Instead, every annotation
processor has to provide that implementation and it can be re-used.