Fix a bug in scope cloning
Merging the passes and depending on AliasAnalysis info (or rather its
scope) revealed what appears to be a long-standing, deeply hidden bug in
scope cloning.
Previously, cloning a scope created a new one that was in no way
associated with the ones present in the graph's chain of scopes. That
meant that various names were not found.
While we managed to somehow avoid this problem, UppercaseNames now
heavily relies on this information being correct.
Notice that for example ChangesetBuilderTest has a weird casting cause
by what appears to be a bug in DemandAnalysis.