Workaround of JS foreign functions definitions
On a scale of 1 to 10, the hack is probably 5. All the alternatives I
considered, involved tweaking IR in a horrible way. And JS is a bit of a
special case TBH.

The solution adds an implicit self parameter with a default value
resolving to the definining module. That way users don't have to provide
an extra  parameter and all arguments are magically applied