temp workaround for the last minute bug
In the current approach the order of calling local functions matters
which came up accidentally in a method with default arguments in
Delimited_Reader. This is because methods are added to the scope in the
order they are processed.
Suprised this hasn't come up before. Will need to address this before
releasing.

Added a (pending) test-case to demonstrate the problem.