JS functions must use this instead of self
This was requested to resolve the exceptional status of `this` in JS.
It also introduces an inconsistency to the language as everything else
uses `self`.
Added some logic to detect situations when someone uses `self` in JS and
issue warnings in such situation.

The implementation implicitly replaces `this` with `self` in the
parameter list so that we don't have to change the whole compiler to
support this exception.