Deal with conflicts
When a virtual module conflicts with a non-virtual module, magic
happens.
We enhance the non-virtual module with essentially exactly the same
information that the virtual one would carry without loosing anything.

In order to distinguish between regular imports and the virtual ones, an
additional flag needs to be carried in order to be able to continue to
do name resolution properly.

Also fixed a long standing problem with imports that include names but
also import the name of the module automatically. Binding map was
incorrectly always taking into account the name of the import.