Force recompilation if imported module has changed
IR cache never really took into account a situation when a binding from
the imported module has changed. In other words, it would continue to
happily use the serialized metadata without noticing that it changed.

This change forces cache invalidation when any of the imported modules
was invalidated (or rather not loaded from cache).
Sadly, it doesn't appear that we have appropriate infrastructure to
write such scenarios so I'm delaying that bit for now.