Entering modules in component browser (#3719)
[ci no changelog needed]

[Task link](https://www.pivotaltracker.com/story/show/181870304).

This PR implements Controller-level support for entering modules in the component browser.

Clicking headers/entries opens the module, filling the Local Scope and Submodules sections.

Please ignore multiple visual issues with the component browser on the screen - this PR does not cause them, and they will be fixed soon when we change the implementation of the component list panel to Grid View.

https://user-images.githubusercontent.com/6566674/190933275-c544ea8d-529d-4b49-820f-d36874344572.mp4

# Important Notes
- We will implement the actual breadcrumbs integration in a separate PR: https://www.pivotaltracker.com/story/show/182675703
- The FRP changes in the component list panel code are temporary. The GridView-based implementation will later replace most of the boilerplate.