Workaround for native image bug until we upgrade
To workaround a bug in abstract classes that manifests itself in sqlite
dependency we have to exclude it from classpath when building the native
image.
Unfortunately assembly plugin does not allow for specifying this
particular setting only for native image generation, therefore had to
define a separate project just for that.
Once we upgrade we can move it back to `engine-runner`.
To invoke native image generation use
`sbt> engine-runner-native/buildNativeImage`