ArrayBranchNode replaced by Array type matching
Polyglot values coming from JS or Python were problematic because they
followed a different path than Java's polyglot values.
The type of JS' Array values was PolyglotProxy which we need to
specialize on.
Python's Array values are on the other hand represented as Python list.
Since it is impossible to special case those when inferring the type,
CatchType has a specialization on Array-like values that have a meta
object but no type.

Additionally added support for pattern matching on polyglot symbols
(both in types and as values).