Split registration of builtins
During Builtins initialization we only register non-static builtins.
That way one does not have to import stdlib to have basic operations on
Any or Number.
Similarly during IRtoTruffle for only consider static builtins, thus
avoiding a situation where we would try to register them twice.

This required changing our annotations processor so that it is aware if
a builtin is a static or not, based on the lack or presence of self
parameter, respectively.

More importantly this approach allows us to avoid hacky solution in
types which caused adding module methods to types' scope. And we don't
do that late in the interpreter.