Normalize index in Vector.at in builtin method
The main culprit of a Vector slowdown (when compared to Array) was the
normalization of the index when accessing the elements.
Turns out that the Graal was very persistent on **not** inlining that
particular fragment and that was degrading the results in benchmarks.

Being unable to force it to do it (looks like a combination of thunk
execution and another layer of indirection) we resorted to just moving
the normalization to the builtin method. That makes Array and Vector
perform roughly the same.