Carry information about the owner of the builtin
Builtin statics are registered differently depending on whether their
owner would be a module or a type. The latter can be registered early on
because we already have type presents, while the former will be
registered during codegen.

This distinction allows us to treat all builtins the same. There is no
need to make a stub definition of a builtin method - they will all be
automatically registered, and the fact if they are static or not is now
irrelevant.