Inject modules' builtin methods defs before passes
Module builtin methods declarations had to be implicitly added at an
early stage so that BindingAnalysis (and later passes) can make use of
that information.
This change adds implicit method definitions for all module builtins
unless they have been provided explicitly.

With this change one does not have to provide explicit builtin
declarations ever, irrespective of whether the owner is a module or a
type. Removed some stub definitions to illustrate the usage.