Commit must force all pending autosaves
Ensured that all commit requests, first ensure that all pending saves to
the files of the project are done. That way, we are not left with
pending file modifications **after** the project was explicitly saved.
Added a test to illustrate the scenario.

Note that one occassion we have to carry around the sender explicitly,
because context changes it for the message once we want to forward it.

Added some actor names, so that it is easier to debug potential issues
with messages.