Change spacing rules to match the autoformatter
There is a mismatch in the spacing rules as described in the style guide and as used by the autoformatter. I've changed the `Spacing` section to follow the autoformatter spacing rules. If this is a desired change, then it probably will be good to apply these rules consistently to all examples in the style guide.