Don't import extension methods by default
Previously, a qualified import would always bring extension methods into
the scope. This is because extensions methods weren't taking into
account the qualifiers at all and imported scope was blindly added to
the existing module scope.

The change limits that by taking into account import qualifiers of the
types that are being brought and creating a copy of the original
`ModuleScope`.

This meant that in pretty much every test file of our testsuite we had
to import `Test` module.
Also, note that one cannot
```
import Standard.Test as Test_Module
```
because of the 2-component name restriction that gets desugared to
`Standard.Test.Main` and we have to write
```
import Standard.Test.Main as Test_Module
```
all over the place.