Ignore imported module in from/all name resolution
It appears that when were doing
`from XYZ import all`
the module `XYZ` was also being taken into account during name
resolution.
This was unfortunate and became problematic when one had a type with the
same name defined in it.
During pattern matching one could not simply do
```
from XYZ import all
...
    case ... of
        _ : XYZ -> ...
```
since the compiler would complain that we try to pattern match on a type
but give it a module.

The module is now excluded from the name resolution, when importing
everything from the module.
It appears that this "feature" was used in a number of our tests,
so they had to be adapted.
This fixes task 4 in https://www.pivotaltracker.com/story/show/183833055