Improvements to VCS setup
The presence of `.git` in Enso's data directory (`.enso`), which should
be under VCS as well, turned out to be really problematic.
Git (and JGit) treat `.enso` as a submodule which prevents as from
manual inspection of the repo. Additionally, every commit changed the
submodule version, which in turn was reporting invalid/changed status
always.

This change renames `.enso/.git` to `.enso/.vcs` which eliminates a lot
of problems. Additionally, rather than doing `git add .` we selectively
add files to the index, thus preventing most of the issues reported in
the ticket.
This change also means that JGit's status needs to filter `.enso/.vcs`
as those will be marked as untracked always.
To manually test the status of the project's repo do
```
git --git-dir .enso/.vcs status
```