Simplify IR for nested patterns
`NestedPatternMatch` pass desugared complex patterns in a very
inefficient way resulting in an exponentional generation of the number
of IR (and Truffle) nodes. Every failed nested pattern would copy all
the remaining patterns of the original case expression.

This change moves the fallthrough logic directly to Truffle nodes. That
way we can generate much simpler IR for nested patterns.

For
```
case x of
    Cons (Cons a b) Nil -> a + b
    Cons a Nil -> a
    _ -> 0
```

we used to generate
```
case x of
    Cons w y -> case w of
        Cons a b -> case y of
            Nil -> a + b
            _ -> case x of
                Cons a z -> case z of
                    Nil -> a
                    _ -> case x of
                        _ -> 0
                _ -> 0
        _ -> case x of
            Cons a z -> case z of
                Nil -> a
                _ -> case x of
                    _ -> 0
           _ -> 0
    Cons a z -> case z of
        Nil -> a
        _ -> case x of
            _ -> 0
    _ -> 0
```
Now it is
```
case x of
    Cons w y -> case w of
        Cons a b -> case y of
            Nil -> a + b
    Cons a z -> case z of
        Nil -> a
    _ -> 0
```