Make ArrayOverBuffer behave like an Array
Most of the problems with accessing `ArrayOverBuffer` have been resolved
by using `CoerceArrayNode` (https://github.com/enso-org/enso/pull/3817).
In `Array.sort` we still however specialized on Array which wasn't
compatible with `ArrayOverBuffer`.

Added a specialization to `Array.sort` to deal with that case. Because
one cannot use a custom comparator with primitive (Java) arrays there is
a conversion needed. It's a necessary penalty if we want to keep
`ArrayOverBuffer` around.

Also fixed an example in `Array.enso` by providing a default argument.