Make Array.sort not sorting in place
Sorting in place would be impractical to support for Arrays coming from
JS or Python.
That is why `Array.sort` now returns a copy of the original Array(-like)
structure, sorted.
This eliminates the need for createing a copy of Array for
`Vector.sort`.

Applied DRY to displaying Array-like structures.