Resolve fully qualified names
Added a separate pass, `FullyQualifiedNames`, that partially resolves
fully qualified names. The pass only resolves the library part of the
name and replaces it with a reference to the `Main` module.

There are 2 scenarios covered in this change:
1) the code uses a fully qualified name to a component that has been
   parsed/compiled
2) the code uses a fully qualified name to a component that has not be
   imported

For the former case, it is sufficient to just check `PackageRepository`
for the presence of the library name. In the latter case we have to
ensure that the library has been already parsed and all its imports are
resolved. Hence the reference to `Compiler` in the `FullyQualifiedNames`
pass.