Avoid IndexOutOfBounds when edits go out of range
While doing regular node manipulation in editior, noticed a number of
situations when

```
java.lang.IndexOutOfBoundsException: None
  at java.base/java.lang.Character.offsetByCodePoints(Character.java:8699)
  at java.base/java.lang.String.offsetByCodePoints(String.java:820)
  at org.enso.text.buffer.CodePointView$Ops$.drop(CodePointView.scala:98)
  at org.enso.text.buffer.CodePointView$Ops$.drop(CodePointView.scala:57)
  at org.enso.text.buffer.Node.drop(Tree.scala:218)
  at org.enso.text.buffer.Rope.dropWith(Rope.scala:86)
  at org.enso.text.buffer.CodePointView.drop(CodePointView.scala:30)
  at org.enso.text.editing.RopeTextEditor$.cutOutTail(RopeTextEditor.scala:42)
...
```
would be thrown. Further text edits would simply be rejected requiring a
complete restart. I doubt we should propagate
`IndexOutOfBoundsException`. Instead it is safer to just apply the edit
to the end of the rope.