Provide better diagnostics when handlers crash
In the event when the action of the underlying actor crashes really bad,
the result of the future will be `Failure`. All such results will be
then wrapped in `Status.Failure` via `pipeTo` (unlike `Success` which
just forwards the response).

In most of the cases we don't handle `Status.Failure` messages, meaning
we are rather left in the dark to the reason of the failure with a
rather non-informative message `Received unknown message: class
akka.actor.Status$Failure`.

We certainly should provide such cases to more handlers but I'm keeping
it small on purpose to keep this change self-contained.