Custom type conversion to double from large long
When a large long would be passed to a host call expecting a double,
it would crash with a
```
Cannot convert '<some long>'(language: Java, type: java.lang.Long) to Java type 'double': Invalid or lossy primitive coercion
```

That is unlikely to be expected by users. It also came up in the
Statistics examples during Sum. One could workaround it by forcing the
conversion manually with `.to_decimal` but it is not a permanent
solution.

Instead this change adds a custom type mapping from Long to Double that
will do it behind the scenes with no user interaction. The mapping kicks
in only for really large longs.