Report type of expressions returning polyglot values
Expressions returning polyglot values were not reporting the type of the
result because we have to do additional magic that infers the correct
Enso type. Since this is exactly what `TypeOfNode` does, I re-used the
logic.

Except this doesn't work in tests because they have assertions on and we
get
```
[enso] WARNING: Execution of function main failed (Invalid library usage. Cached library must be adopted by a RootNode before it is executed.).
java.lang.AssertionError: Invalid library usage. Cached library must be adopted by a RootNode before it is executed.
```