JGit should run its bookkeeping in the background
JGit ops generally run fast (as in a few milliseconds) except for the
first commit. The initialization + first commit was taking at least 3.5
seconds constistenly, but only in the first test case.
Now, this led to frequent timeouts down the chain when the request was
expected to finish fast.
The bookkeeping involved some timestamping and other expensive calls in
order to calculate clock drift. The default appears to be to run it in a
blocking mode, hence adding at least 3 seconds to the first command
call.

Setting the job to run in the background makes the cost of the repo
initialization, acceptable (200-400 milliseconds on a cold JVM). The
other commands are unaffected and take < 10 milliseconds.