Stabilize flaky test that re-uses test setup
Follow up on
https://github.com/enso-org/enso/pull/4072/commits/16ba57d4652d26ecd74b6f2b212b96773a0e7145
which marked the tests as flaky due to the discovered race-condition
when cleaning up the test context.

This test goes back to the original implementation where JGit did all
its book-keeping synchronously. When we switched to asynchronous
operation, cleaning up the test directory would sometimes not succeed
because there would be a race-condition with the ongoing work of JGit.

Additionally, in the current setup tagging a test as flaky is only
valid for Windows. The other platforms ignore the flag. So we were still
getting random failures.