Fix serialization of FQNs
Serialization of FQNs' metadata was broken - we attempted to serialiaze
concrete modules and that's against the design.

Added a test illustrating the problem which would previously fail during
serialization.

It is still hard to discover problems like this because
SerializationManager creates system threads; when the exception occurs
there it is typically not the real cause. Creating regular threads via
`createThread` seems to be problematic for `native-image`.
There is a simple workaround for the former but will leave it out for
another PR to simplify the review.