wip: Save and load library bindings
This change adds serialization and deserialization of library bindings.
In order to be functional, one can needs to first generate IR and
serialize bindings `--compiled <path-to-library>` command. The bindings
will be stored under the library with `.bindings` suffix.

When resolving import/exports the compiler will first try to load
module's bindings from cache. If successful, it will not schedule its
imports/exports for immediate compilation, as we always did, but use the
bindings info to infer the dependent modules.

The current change does not make any optimizations when it comes to
compiling the modules, yet. It only delays the actual
compilation/loading IR from cache so that it can be done in bulk.
Further optimizations will come from this opportunity such as parallel
loading of caches or lazilly inferring only the necessary modules.

Other important things to note:
 - deserialization has to take into account dependencies of modules,
   otherwise one may not be able to fully restore abstract module
references
 - bindings map must have a stable serialization uid in order to avoid
   spurious and costly failed cache loading attempts