Fix unhandled case in equality
Seems like my code managed to trigger a case when comparing an
`AtomConstructor` wrapped in a `Warning` with another
`AtomConstructor`.

Somehow we are already past `equalsWithWarnings` and enter the fallback
specialization.
Not yet sure how to reproduce this in a single test but my weird failing
test in `Integrations_Tests` now succeeds.