Downgrade hashing to SHA-1 and other optimizations
This change downgrades hashing algorithm used in caching IR and library
bindings to SHA-1. It is sufficient and significantly faster for the
purpose of simple checksum we use it for.

Additionally, don't calculate the digest for serialized bytes - if we
get the expected object type then we are confident about the integrity.

Don't initiallize Jackson's ObjectMapper for every metadata
serialization/de-serialization. Initialization is very costly.

Avoid unnecessary conversions between Scala and Java. Those
back-and-forth `asScala` and `asJava` are pretty expensive.

Finally fix a warning when generating library cache.