Don't log installed engines and runtimes in prod
The action will check and install all necessary components of GraalVM
for every mentioned version. While not harmful, it adds up to startup
time.
Additionally added an option in language server startup to skip
installation of GraalVM components. The latter is already performed by
project-manager when opening the project and it is unnecessary to do it
twice. Due to LS' architecture this configuration has to be passed
around via multiple configs.

Finally, skipped the attempt to install Python component on Windows -
this is not supported by GraalVM atm.

Closes #5749.