Detect conflicts between exported types and FQNs
Exporting types named the same as the module where they are defined in
`Main` modules of library components may lead to accidental name
conflicts. This became apparent when trying to access `Problem_Behavior`
module via a fully qualified name and the compiler rejected it.
This is due to the fact that `Main` module exported `Error` type defined
in `Standard.Base.Error` module, thus making it impossible to access any
other submodules of `Standard.Base.Error` via a fully qualified name.

This change adds a warning to FullyQualifiedNames pass that detects any
such future problems.
While only `Error` module was affected, it was widely used in the
stdlib, hence the number of changes.

I left out the potential conflict in micro-distribution, thus ensuring
we actually detect and report the warning.

Closes #5902.