Report errors when generating indexes
`--compile` command would run the compilation pipeline but silently
omit any encountered errors, thus skipping the serialization.
This maybe was a good idea in the past but it was problematic now that
we generate indexes on build time.
This resulted in rather obscure errors (#6092) for modules that were
missing their caches.

Making compilation more resilient to sudden cache misses is a separate
item.