Engine benchmark visualization tool can compare two bench runs (#6198)
Add `--compare <bench-run-id-1> <bench-run-id-2>` option to `bench_download.py` script, that prints the difference of benchmark scores for two benchmark runs from GH as a table.

# Important Notes
I tried to add an option to compare two or more branches and visualize them in the graphs, but gave that up after some struggling.