Introduce Warning-related methods to Any
This change adds a special method dispatch to values with warnings.
That way one can call methods like `has_warnings`, `remove_warnings` or
`throw_on_warning` on any value and get the expected results.

The change requires a special handling in method dispatch because of the
way warnings are being propagated during execution - we unwrap and then
append warnings during method calls.
One additional problem here was that methods on Warning builtin type are
part of the eigentype and have a synthetic `self` parameter, which would
make calls **not** fully applied unless we add that (useless but still)
argument during method dispatch.