UnresolvedSymbol is now accepted by Vector.sort
`Vector.sort` does some custom method dispatch logic which always
expected a function as `by` and `on` arguments. At the same time,
`UnresolvedSymbol` is treated like a (to be resolved) `Function` and
under normal circumstances there would be no difference between
`_.foo` and `.foo` provided as arguments.

Rather than adding an additional phase that does some form of
eta-expansion, to accomodate for this custom dispatch, this change only
fixes the problem locally. We accept `Function` and `UnresolvedSymbol`
and perform the resolution on the fly. Ideally, we would have a
specialization on the latter but again, it would be dependent on the
contents of the `Vector` so unclear if that is better.