Report only unique warnings
This change makes sure that reported warnings are unique, based on the
value of internal clock tick and ignoring differences in reassignments.

The `Warning.get_all` is the main entry point to all requests for
values' warnings. So rather than implementing de-duplication all over
the place, we just do it in one place. There is a cost involved in
deduplication (sorting, copying the array + going over the whole array)
but that seems like an acceptable cost.