Fix scrolling issue
The core of the issue was that the length the scrollbars scroll over
didn't take padding into account. At the same time, I changed the `max`
and `thumb_size` variables to the coordinate system of the content. This
is also how they're being used in `ScrollArea`, which is the only other
place where `Scrollbar`s are being used.